replaced goal amount by double type

+++ analysis/src/main/java/com/savik/football/bets/goals/match/AtLeastOneTeamNotScoredAndUnder.java
@@ -10,7 +10,7 @@ public class AtLeastOneTeamNotScoredAndUnder extends GeneralBet {
 
     private MatchUnder matchUnder;
 
-    public AtLeastOneTeamNotScoredAndUnder(int amount) {
+    public AtLeastOneTeamNotScoredAndUnder(double amount) {
         bothTeamsScored = new BothTeamsScored();
         matchUnder = new MatchUnder(amount);
     }
