+++ analysis/src/main/java/com/savik/football/bets/total/over/match/MatchHomeOver.java
@@ -4,9 +4,9 @@ import com.savik.football.bets.GeneralBet;
 import com.savik.football.model.Match;
 
 public class MatchHomeOver extends GeneralBet {
-    private int amount;
+    private double amount;
 
-    public MatchHomeOver(int amount) {
+    public MatchHomeOver(double amount) {
         this.amount = amount;
     }
