+++ analysis/src/main/java/com/savik/football/bets/total/under/second_period/SecondPeriodLoserUnder.java
@@ -6,9 +6,9 @@ import com.savik.football.model.Who;
 
 public class SecondPeriodLoserUnder extends GeneralBet {
 
-    private int amount;
+    private double amount;
 
-    public SecondPeriodLoserUnder(int amount) {
+    public SecondPeriodLoserUnder(double amount) {
         this.amount = amount;
     }
